+++ niveshpy/services/parsing.py
@@ -3,7 +3,7 @@
 from collections.abc import Callable, Sequence
 from typing import TypeVar
 
-from attrs import field, frozen
+from attrs import evolve, field, frozen
 
 from niveshpy.domain.repositories import (
     AccountRepository,
@@ -53,7 +53,10 @@ class ParsingService:
     def _add_metadata(self, item: _T) -> _T:
         """Add metadata to a parsed item."""
         if item.properties.get("source") is None:
-            item.properties["source"] = "parser"
+            if isinstance(item, (AccountCreate, SecurityCreate)):
+                item = evolve(item, properties={**item.properties, "source": "parser"})
+            else:
+                item.properties["source"] = "parser"
         return item
 
     def _bulk_insert_accounts(
