+++ niveshpy/services/price.py
@@ -25,7 +25,7 @@ from niveshpy.exceptions import (
 from niveshpy.models.output import BaseMessage, ProgressUpdate, Warning
 from niveshpy.models.price import PriceCreate, PricePublicWithRelations
 from niveshpy.models.provider import Provider, ProviderInfo
-from niveshpy.models.security import Security
+from niveshpy.models.security import SecurityPublic
 
 
 @dataclass(slots=True, frozen=True)
@@ -159,7 +159,7 @@ class PriceService:
 
         return provider_instances
 
-    def _fetch_securities(self, queries: tuple[str, ...]) -> Sequence[Security]:
+    def _fetch_securities(self, queries: tuple[str, ...]) -> Sequence[SecurityPublic]:
         """Fetch securities matching the given queries.
 
         Args:
@@ -183,10 +183,10 @@ class PriceService:
 
     def _build_provider_map(
         self,
-        securities: Sequence[Security],
+        securities: Sequence[SecurityPublic],
         provider_instances: list[tuple[str, ProviderInfo, Provider]],
     ) -> tuple[
-        dict[str, list[tuple[int, str, ProviderInfo, Provider]]], list[Security]
+        dict[str, list[tuple[int, str, ProviderInfo, Provider]]], list[SecurityPublic]
     ]:
         """Build provider priority map for each security.
 
@@ -202,7 +202,7 @@ class PriceService:
             Tuple of (provider_map, securities_with_providers).
         """
         provider_map: dict[str, list[tuple[int, str, ProviderInfo, Provider]]] = {}
-        securities_with_providers: list[Security] = []
+        securities_with_providers: list[SecurityPublic] = []
 
         for security in securities:
             applicable_providers: list[tuple[int, str, ProviderInfo, Provider]] = []
@@ -222,7 +222,7 @@ class PriceService:
 
     def _fetch_prices_for_security(
         self,
-        security: Security,
+        security: SecurityPublic,
         providers: list[tuple[int, str, ProviderInfo, Provider]],
         start_date: datetime.date,
         end_date: datetime.date,
@@ -437,12 +437,12 @@ class PriceService:
 
     def _process_sync(
         self,
-        security: Security,
+        security: SecurityPublic,
         providers: list[tuple[int, str, ProviderInfo, Provider]],
         force: bool,
     ) -> (
         tuple[
-            Security,
+            SecurityPublic,
             list[PriceCreate],
             str,
             datetime.date,
